requirements.txt
fastapi
duckdb
pandas
openpyxl
langchain-core
langchain-community
langchain-gigachat
python-dotenv
numpy
scikit-learn
uvicorn